refactor: move two methods to MasterTable
move methods `alignOverlaysWithTrimmingContainer`, `markOversizedColumnHeaders` to MasterTable,
because they were no-ops in other (overlay) tables
