+++ src/3rdparty/walkontable/src/table/_base.js
@@ -1,13 +1,7 @@
 import Table from '../table';
 
 class OverlayTable extends Table {
-  alignOverlaysWithTrimmingContainer() {
 
-  }
-
-  markOversizedColumnHeaders() {
-
-  }
 }
 
 export default OverlayTable;
